Part 2 timer functionality working.
There exists a timing::Timer object that manages the time state for each board. The timing::Timer::run() method is run on its own thread and fires an event when the Timer object's counter is incremented. The controller then propagates this event back to the view object which then updates the Fl_Output that displays the timer information.