Game time values are now saved.
The amount of time that it takes for a player to complete a level is saved when they close the game. I think this is important to prevent high-score cheating. If they were not saved, for example, then a player could take as long as they wanted to complete the problem, then right before they finish, they could close the game and reopen it to submit the finished solution with a timer of only 0 or 1 second.