The game can now be paused.
All tile inputs are hidden when the game is paused and the timer is stopped if it is running. The pause button now functions as a toggle. Clicking it multiple times now swaps between the paused/unpaused state. I also increased the size of the game so we could have more room as it started to look a bit squished.