Fixed reset button not working.
With the recent change to not allow duplicate values. The reset was not working as intended because GameController::resetCurrentPuzzle() was attempting to set all values to 0; however, the new changes did not allow duplicate values. Therefore, I just added an exception for 0 in TileGame::trySetTileValue(int, int) and it is working fine now.